requirements.txt
torch>=1.10.0
transformers>=4.18.0
gradio>=3.0.0
Pillow>=8.0.0